chore: fix iOS analytics
